Fill out 'submit' command
Users can now submit a problem based on it's name rather than it's PID
(the PID is now optional). In addition, the ability to specify python
version was added. Finally, if the submission was successful, the user
is prompted with thre current status of the problem at one second
intervals until the judging is complete.