Create competition subcommand
The `competition` subcomman allows for lookup of all or one competition.
If no competition is specified, general data about all competitions is
presented. If a competition is specified (currently only competition id
lookup is supported), that specific competition is looked up and
detailed info about that competition is presented (users, problems,
etc.)